tried to repair PostProcessing2D
it seems it now works now to set 'output_write_case' and 'output_write_vtk'
independently. This was not the case before. Also, writing output only every
'steps_per_output' is not possible.

This class is not tested, so I hope I didn't brake anything.

+++ src/General/ParameterDatabase.C
@@ -1380,7 +1380,7 @@ ParameterDatabase ParameterDatabase::default_output_database()
 	         "This string is prepended to most files written by ParMooN. "
 	         "This includes also vtk- and case-files");
 
-	  db.add("steps_per_output", 1,
+	  db.add("steps_per_output", (size_t)1,
 	         "This integer specifies how many (time) steps are performed "
 		 "before writing the results ");
